+++ src/agent/samplers/gpu/linux/nvidia/stats.rs
@@ -138,6 +138,27 @@ pub static GPU_DRAM_BW_UTILIZATION: GaugeGroup = GaugeGroup::new(MAX_GPUS);
 #[metric(
     name = "gpu_tensor_utilization",
     description = "The percentage of time the GPU's SMs were doing any tensor operations. (0-100). Requires Hopper+ GPU.",
-    metadata = { unit = "percentage" }
+    metadata = { precision = "any", unit = "percentage" }
 )]
 pub static GPU_TENSOR_UTILIZATION: GaugeGroup = GaugeGroup::new(MAX_GPUS);
+
+#[metric(
+    name = "gpu_tensor_utilization",
+    description = "The percentage of time the GPU's SMs were doing HMMA (FP16/BF16) tensor operations. (0-100). Requires Hopper+ GPU.",
+    metadata = { precision = "fp16", unit = "percentage" }
+)]
+pub static GPU_TENSOR_UTILIZATION_FP16: GaugeGroup = GaugeGroup::new(MAX_GPUS);
+
+#[metric(
+    name = "gpu_tensor_utilization",
+    description = "The percentage of time the GPU's SMs were doing IMMA (integer) tensor operations. (0-100). Requires Hopper+ GPU.",
+    metadata = { precision = "int8", unit = "percentage" }
+)]
+pub static GPU_TENSOR_UTILIZATION_INT8: GaugeGroup = GaugeGroup::new(MAX_GPUS);
+
+#[metric(
+    name = "gpu_tensor_utilization",
+    description = "The percentage of time the GPU's SMs were doing DFMA (FP64) tensor operations. (0-100). Requires Hopper+ GPU.",
+    metadata = { precision = "fp64", unit = "percentage" }
+)]
+pub static GPU_TENSOR_UTILIZATION_FP64: GaugeGroup = GaugeGroup::new(MAX_GPUS);